Separate module in agent and manager versions
