+++ etc/templates/config/generic/wodle-agent-upgrade.template
