+++ src/config/wmodules-config.c
@@ -128,6 +128,11 @@ int Read_WModule(const OS_XML *xml, xml_node *node, void *d1, void *d2)
 #else
         mwarn("The '%s' module is not available on Windows systems. Ignoring it.", node->values[0]);
 #endif
+    } else if (!strcmp(node->values[0], WM_AGENT_UPGRADE_CONTEXT.name)) {
+        if (wm_agent_upgrade_read(children, cur_wmodule) < 0) {
+            OS_ClearNode(children);
+            return OS_INVALID;
+        }
     }
 #ifndef WIN32
 #ifndef CLIENT
@@ -144,11 +149,6 @@ int Read_WModule(const OS_XML *xml, xml_node *node, void *d1, void *d2)
         }
     } else if (!strcmp(node->values[0], WM_KEY_REQUEST_CONTEXT.name)) {
         if (wm_key_request_read(children, cur_wmodule) < 0) {
-            OS_ClearNode(children);
-            return OS_INVALID;
-        }
-    } else if (!strcmp(node->values[0], WM_AGENT_UPGRADE_CONTEXT.name)) {
-        if (wm_agent_upgrade_read(children, cur_wmodule) < 0) {
             OS_ClearNode(children);
             return OS_INVALID;
         }
