+++ src/init/inst-functions.sh
@@ -35,7 +35,7 @@ CISCAT_TEMPLATE="./etc/templates/config/generic/wodle-ciscat.template"
 VULN_TEMPLATE="./etc/templates/config/generic/wodle-vulnerability-detector.manager.template"
 
 SECURITY_CONFIGURATION_ASSESSMENT_TEMPLATE="./etc/templates/config/generic/sca.template"
-AGENT_UPGRADE_TEMPLATE="./etc/templates/config/generic/wodle-agent-upgrade.manager.template"
+AGENT_UPGRADE_TEMPLATE="./etc/templates/config/generic/wodle-agent-upgrade.template"
 TASK_MANAGER_TEMPLATE="./etc/templates/config/generic/wodle-task-manager.manager.template"
 
 ##########
@@ -389,6 +389,10 @@ WriteAgent()
 
     # Syscollector configuration
     WriteSyscollector "agent"
+
+    # Agent upgrade
+    cat ${AGENT_UPGRADE_TEMPLATE} >> $NEWCONFIG
+    echo "" >> $NEWCONFIG
 
     # Configuration assessment configuration
     WriteConfigurationAssessment
