+++ src/wazuh_modules/agent_upgrade/manager/wm_agent_upgrade_commands.c
@@ -9,7 +9,7 @@
  * Foundation.
  */
 #include "wazuh_modules/wmodules.h"
-#include "wm_agent_upgrade.h"
+#include "wm_agent_upgrade_manager.h"
 #include "wm_agent_upgrade_parsing.h"
 #include "wm_agent_upgrade_tasks.h"
 #include "os_net/os_net.h"
