+++ src/wazuh_modules/agent_upgrade/manager/wm_agent_upgrade_tasks.c
