+++ src/win32/ossec-pre6.conf
@@ -136,6 +136,11 @@
     </synchronization>
   </syscheck>
 
+  <!-- Agent upgrade -->
+  <wodle name="agent-upgrade">
+    <enabled>yes</enabled>
+  </wodle>
+
   <!-- System inventory -->
   <wodle name="syscollector">
     <disabled>no</disabled>
