+++ src/win32/ossec.conf
@@ -155,6 +155,11 @@
     </synchronization>
   </syscheck>
 
+  <!-- Agent upgrade -->
+  <wodle name="agent-upgrade">
+    <enabled>yes</enabled>
+  </wodle>
+  
   <!-- System inventory -->
   <wodle name="syscollector">
     <disabled>no</disabled>
